refactor(testpage): make LiveNavTestPage partial and split it by surface
Seven surfaces out of the 2,746-line class: parts, actions, lifecycle, write
buffer, fields, navigation, row addressing. One line of the 4,638 changes --
'internal class' becomes 'internal partial class'. 41 build warnings before and after.
